refactor engines into a folder
