+++ marimo/_runtime/runner/hooks_post_execution.py
@@ -32,7 +32,10 @@ from marimo._runtime.context.types import get_context, get_global_context
 from marimo._runtime.control_flow import MarimoInterrupt, MarimoStopError
 from marimo._runtime.runner import cell_runner
 from marimo._server.model import SessionMode
-from marimo._sql.engines import INTERNAL_DUCKDB_ENGINE, DuckDBEngine
+from marimo._sql.engines.duckdb import (
+    INTERNAL_DUCKDB_ENGINE,
+    DuckDBEngine,
+)
 from marimo._sql.get_engines import (
     engine_to_data_source_connection,
     get_engines_from_variables,
