+++ tests/_sql/test_engines.py
@@ -7,9 +7,9 @@ from unittest.mock import MagicMock
 import pytest
 
 from marimo._dependencies.dependencies import DependencyManager
-from marimo._sql.clickhouse_engines import ClickhouseEmbedded
-from marimo._sql.engines import (
-    DuckDBEngine,
+from marimo._sql.engines.clickhouse import ClickhouseEmbedded
+from marimo._sql.engines.duckdb import DuckDBEngine
+from marimo._sql.engines.sqlalchemy import (
     SQLAlchemyEngine,
     _sql_type_to_data_type,
     raise_df_import_error,
